refactor(engine): move segment flushers + GPU plumbing into GpuReplay (T10c)
Third sub-PR of T10 (replay/submit split). Moves onto GpuReplay: the 5 replay-only
GPU-plumbing fields (viewport_buffer, viewport_bind_group, unit_quad_buffer,
unit_quad_index_buffer, default_sampler) and the 6 segment flushers (flush_segment,
flush_all_instanced_batches, flush_gradient_batches, flush_tessellated_geometry,
flush_segment_cached_images, flush_segment_external_images). The T10b
flush_texture_batch family drops its borrowed-plumbing params (now self-fields).

GpuReplay::new(&device, &pipelines, w, h) builds the fields; the viewport_bind_group
is created against PipelineSet's shared layout (wgpu object identity preserved).
resize() delegates to replay.update_viewport (byte-identical buffer write). The old
flush_segment swap-trick is correctly eliminated — flushers now take &mut DrawSegment
directly. painter render()/flush_opacity_layer stay (T10d), calling
self.replay.flush_segment(seg, self.size, &device, &queue, &mut pipelines, &mut
resources, encoder, view). painter.rs 5073 -> 4372; replay.rs 302 -> 1121.

R1 (the fixed 5-phase flush order instanced->gradient->tessellated->cached->external)
moves as an atomic unit, byte-identical.

Gates: clippy both modes, fmt, taplo, typos, doc, port-check green. GPU-readback
serial (DX12) 216/0/7 (pure move). Reviewed by rust-reviewer + ce-kieran (both
COMPLETE, no findings — R1 order, swap-elimination, viewport layout identity, resize
byte-identity, viewport_size equivalence, and the borrow seam all verified vs main +
the readback oracle).
